Fix bug by converting all vals in tensor to floats

+++ src/types.hpp
@@ -437,18 +437,18 @@ namespace madrona_gpudrive
 
     struct MetaData
     {
-        int32_t isSdc;
-        int32_t isObjectOfInterest;
-        int32_t isTrackToPredict;
-        int32_t difficulty;
+        float isSdc;                
+        float isObjectOfInterest;   
+        float isTrackToPredict;     
+        float difficulty;           
         float avgZ;
-
+    
         static inline void zero(MetaData& metadata)
         {
-            metadata.isSdc = -1;
-            metadata.isObjectOfInterest = -1;
-            metadata.isTrackToPredict = -1;
-            metadata.difficulty = -1;
+            metadata.isSdc = -1.0f;           
+            metadata.isObjectOfInterest = -1.0f; 
+            metadata.isTrackToPredict = -1.0f;   
+            metadata.difficulty = -1.0f;         
             metadata.avgZ = 0.0f;
         }
     };
